build(webpack): Support for webpack MultiCompiler
Webpack supports multiple configurations under one runtime, which is helpful in particular with Server-Side Rendering.
Previously the NestJS CLI build command wasn't prepared for this configuration in webpack.config.js and errored out with
an non-descript error. With this commit, the NestJS CLI build command properly supports this configuration mode.

<footer>